requirements.txt
scrapy
django-bower
django-pipeline
django-autoslug
unidecode
django-autocomplete-light
